+++ platforms/gradle/foundry-gradle-plugin/src/main/kotlin/foundry/gradle/FoundryProperties.kt
@@ -253,7 +253,7 @@ internal constructor(
 
   /** Opt out for -Werror in tests. */
   public val allowWarningsInTests: Provider<Boolean>
-    get() = resolver.booleanProvider("foundry.kotlin.allowWarningsInTests", defaultValue = false)
+    get() = resolver.booleanProvider("foundry.kotlin.allowWarningsInTests", defaultValue = allowWarnings)
 
   /** Opt-in for explicit API mode. Maps to [ExplicitApiMode] */
   public val kotlinExplicitApiMode: Provider<String>