+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changelog
 - **Change**: `foundry.kotlin.allowWarningsInTests` now defaults to the value of `foundry.kotlin.allowWarnings`.
 - **New**: Add a `foundry.android.test.compressWithLegacyPackaging` flag to compress androidTest APKs with legacy packaging.
 - **New**: Support emulator.wtf for `androidTest()`. This feature is gated by the `foundry.emulatorwtf.enable` feature flag.
+- **Enhancement**: Module topography validation now supports custom replacements for matching regexes.
 - Update to Kotlin `2.1.0`.
 - Build against KSP `2.1.0-1.0.29`