+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 **Unreleased**
 --------------
 
+0.23.0
+------
+
+_2024-12-10_
+
 - **New**: Introduce `foundry.compose.global.stabilityConfigurationPath` to specify a global compose stability configuration file.
 - **New**: Introduce `foundry.kotlin.explicitApiMode` to specify an explicit API mode for kotlin projects. Default is null. Values correspond to the `ExplicitApiMode` enum in KGP.
 - **Change**: `foundry.compose.stabilityConfigurationPath` should now be relative from the _project_'s directory, not the root project.
@@ -13,7 +18,15 @@ Changelog
 - **New**: Support emulator.wtf for `androidTest()`. This feature is gated by the `foundry.emulatorwtf.enable` feature flag.
 - **Enhancement**: Module topography validation now supports custom replacements for matching regexes.
 - Update to Kotlin `2.1.0`.
-- Build against KSP `2.1.0-1.0.29`
+- Update to Clikt `5.0.2`.
+- Update to Kaml `0.66.0`.
+- Update to RxJava `3.1.10`.
+- Build against Gradle `8.11.1`.
+- Build against KSP `2.1.0-1.0.29`.
+- Build against Roborazzi `1.36.0`.
+- Build against DAGP `2.6.0`.
+- Build against AGP `8.7.3`.
+- Build against redacted compiler plugin `1.11.0`.
 
 0.22.6
 ------